refactor(client engine): move QueryPlan related types and QueryInterpreter next to ClientEngine
