+++ packages/client/src/runtime/core/engines/common/Engine.ts
@@ -91,104 +91,6 @@ export type RequestBatchOptions<InteractiveTransactionPayload> = {
 
 export type BatchQueryEngineResult<T> = QueryEngineResultData<T> | Error
 
-export type PrismaValuePlaceholder = { prisma__type: 'param'; prisma__value: { name: string; type: string } }
-
-export function isPrismaValuePlaceholder(value: unknown): value is PrismaValuePlaceholder {
-  return typeof value === 'object' && value !== null && value['prisma__type'] === 'param'
-}
-
-export type PrismaValue =
-  | string
-  | boolean
-  | number
-  | PrismaValue[]
-  | null
-  | Record<string, unknown>
-  | PrismaValuePlaceholder
-
-export type QueryPlanBinding = {
-  name: string
-  expr: QueryPlanNode
-}
-
-export type QueryPlanDbQuery = {
-  query: string
-  params: PrismaValue[]
-}
-
-export type JoinExpression = {
-  child: QueryPlanNode
-  on: [left: string, right: string][]
-  parentField: string
-}
-
-export type QueryPlanNode =
-  | {
-      type: 'seq'
-      args: QueryPlanNode[]
-    }
-  | {
-      type: 'get'
-      args: {
-        name: string
-      }
-    }
-  | {
-      type: 'let'
-      args: {
-        bindings: QueryPlanBinding[]
-        expr: QueryPlanNode
-      }
-    }
-  | {
-      type: 'getFirstNonEmpty'
-      args: {
-        names: string[]
-      }
-    }
-  | {
-      type: 'query'
-      args: QueryPlanDbQuery
-    }
-  | {
-      type: 'execute'
-      args: QueryPlanDbQuery
-    }
-  | {
-      type: 'reverse'
-      args: QueryPlanNode
-    }
-  | {
-      type: 'sum'
-      args: QueryPlanNode[]
-    }
-  | {
-      type: 'concat'
-      args: QueryPlanNode[]
-    }
-  | {
-      type: 'unique'
-      args: QueryPlanNode
-    }
-  | {
-      type: 'required'
-      args: QueryPlanNode
-    }
-  | {
-      type: 'join'
-      args: {
-        parent: QueryPlanNode
-        children: JoinExpression[]
-      }
-    }
-  | {
-      type: 'mapField'
-      args: {
-        field: string
-        records: QueryPlanNode
-      }
-    }
-
 export interface Engine<InteractiveTransactionPayload = unknown> {
   /** The name of the engine. This is meant to be consumed externally */
   readonly name: string
